+++ CHANGELOG.md
@@ -7,6 +7,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+### Added
+
+- GPS compass heading using `iio-sensor-proxy`
+- NMEA modem GPS format support
+
 ## 1.5.0 - 2026-02-08
 
 ### Added